extracet class TeXDoc
